+++ contracts/mixins/StorageLayout.sol
@@ -66,7 +66,7 @@ contract StorageLayout {
 
     /**************************************************************/
     // PoolRegistry
-    mapping(uint24 => PoolSpecs.Pool) internal templates_;
+    mapping(uint256 => PoolSpecs.Pool) internal templates_;
     mapping(bytes32 => PoolSpecs.Pool) internal pools_;
     mapping(address => PriceGrid.ImproveSettings) internal improves_;
     uint128 internal newPoolLiq_;
